Se añadieron excepeciones y solucionaron error

+++ Learnia/Learnia/src/main/java/com/proyecto/Learnia/service/CategoriaServiceImp.java
@@ -1,6 +1,8 @@
 package com.proyecto.Learnia.service;
 
 import com.proyecto.Learnia.entity.Categoria;
+import com.proyecto.Learnia.exception.ResourceNotFoundException;
+import com.proyecto.Learnia.exception.SuccesException;
 import com.proyecto.Learnia.repository.CategoriaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -40,6 +42,10 @@ public class CategoriaServiceImp implements CategoriaService {
 
     @Override
     public void eliminar(Long id) {
+        if(!categoriaRepository.existsById(id)){
+            throw new ResourceNotFoundException("La categoria con id: " + id + " no fue encontrado");
+        }
         categoriaRepository.deleteById(id);
+        throw new SuccesException("Categoria eliminada correctamente");
     }
 }
